+++ abidance/exchange/__init__.py
@@ -6,13 +6,16 @@ interacting with various cryptocurrency exchanges.
 """
 
 # Import key classes to make them available at the module level
-from .base import Exchange
+from .base import Exchange as ExchangeBase
 from .manager import ExchangeManager
 from .binance import BinanceExchange
+from .protocols import Exchange, ExchangeFactory
 
 # Define what's available when doing "from abidance.exchange import *"
 __all__ = [
-    "Exchange",
+    "Exchange",  # Protocol
+    "ExchangeBase",  # Abstract base class
+    "ExchangeFactory",  # Protocol
     "ExchangeManager",
     "BinanceExchange",
 ]